+++ thereadingmachine/keyword.py
@@ -1,5 +1,6 @@
 from nltk.corpus import wordnet as wn
 import re
+import json
 
 # The topic and key word functions assumes that we know the topics
 # already. For example, wheat, rice and soybean related topic.
@@ -68,3 +69,62 @@ def extract_text_keywords(text, topic):
                             for word in keywords
                             if word in text])
     return {topic: located_keywords}
+
+
+def tag_commodity(article, wheat_keywords, rice_keywords,
+                  maize_keywords, soybean_keywords,
+                  grain_keywords):
+    keyword = {'hasWheat': 0,
+               'hasRice': 0,
+               'hasMaize': 0,
+               'hasSoybean': 0,
+               'hasGrain': 0
+               }
+
+    wheat_match = set(wheat_keywords).intersection(
+        article['processed_article'])
+    rice_match = set(rice_keywords).intersection(
+        article['processed_article'])
+    maize_match = set(maize_keywords).intersection(
+        article['processed_article'])
+    soybean_match = set(soybean_keywords).intersection(
+        article['processed_article'])
+    grain_match = set(grain_keywords).intersection(
+        article['processed_article'])
+
+    if len(wheat_match) > 0:
+        keyword['hasWheat'] = 1
+    if len(rice_match) > 0:
+        keyword['hasRice'] = 1
+    if len(maize_match) > 0:
+        keyword['hasMaize'] = 1
+    if len(soybean_match) > 0:
+        keyword['hasSoybean'] = 1
+    if len(grain_match) > 0:
+        keyword['hasGrain'] = 1
+    return keyword
+
+
+def get_amis_topic_keywords():
+    # Pre define the topics
+    topics = ['wheat', 'rice', 'maize', 'barley', 'soybean']
+    wheat_keywords, rice_keywords, maize_keywords, barley_keywords, soybean_keywords = [
+        get_topic_keywords(topic) for topic in topics]
+    grain_keywords = list(set(get_topic_keywords('grains')) -
+                          set(wheat_keywords) -
+                          set(rice_keywords) -
+                          set(maize_keywords) -
+                          set(soybean_keywords) -
+                          set(barley_keywords))
+    # Manually remove keywords
+    for index, grain in enumerate(grain_keywords):
+        for topic, topic_keyword in zip(topics,
+                                        [wheat_keywords, rice_keywords,
+                                         maize_keywords,
+                                         soybean_keywords]):
+            search_string = r'\b' + re.escape(topic) + r'\b'
+            if bool(re.search(search_string, grain)):
+                print('popping out ' + (grain))
+                topic_keyword.append(grain_keywords.pop(index))
+
+    return wheat_keywords, rice_keywords, maize_keywords, barley_keywords, soybean_keywords, grains_keywords
